+++ node/.env
@@ -1,0 +1,1 @@
+MONGO_URL='mongodb://localhost/echo'
