+++ node/src/app.js
@@ -17,8 +17,8 @@ db.once('open', function () {
 
 require('./models/User');
 require('./models/Team');
-const User = mongoose.model('User');
-const Team = mongoose.model('Team');
+require('./models/Meeting');
+require('./models/Task');
 
 
 // express
@@ -27,6 +27,8 @@ let express = require("express");
 let bodyParser = require('body-parser');
 let app = express();
 
+app.use(require('./routes'));
+
 app.use(bodyParser.json()); // for parsing application/json
 app.use(bodyParser.urlencoded({ extended: true })); // for parsing application/x-www-form-urlencoded
 
@@ -34,26 +36,6 @@ app.use(function (req, res, next) {
   res.header("Access-Control-Allow-Origin", "*");
   res.header("Access-Control-Allow-Headers", "Origin, X-Requested-With, Content-Type, Accept");
   next();
-});
-
-app.get("/user/:name/:password", async (req, res) => {
-  const newuser = new User({
-    name: req.params['name'],
-    password: req.params['password']
-  });
-  newuser.save(function (e) {
-    if (e) {
-      console.log(e);
-    }
-  });
-  res.json('success');
-});
-
-app.post("/team", async (req, res) => {
-  // const name = req.params['name'];
-  console.log('team post');
-  console.log(req.body);
-  res.json(req.body);
 });
 
 app.listen(3000, () => {
