+++ node/src/routes/api/index.js
@@ -1,0 +1,22 @@
+const router = require('express').Router();
+
+router.use('/user', require('./user'));
+// router.use('/meeting', require('./meeting'));
+// router.use('/task', require('./task'));
+router.use('/team', require('./team'));
+
+router.use(function(err, req, res, next){
+  if(err.name === 'ValidationError'){
+    return res.status(422).json({
+      errors: Object.keys(err.errors).reduce(function(errors, key){
+        errors[key] = err.errors[key].message;
+
+        return errors;
+      }, {})
+    });
+  }
+
+  return next(err);
+});
+
+module.exports = router;
